feat: add icons to sliders and allow mouse sensitivity and scroll speed values below 1

+++ app/src/main/java/xtr/keymapper/server/IInputInterface.java
@@ -9,7 +9,7 @@ import xtr.keymapper.server.event.MouseEventHandler;
 public interface IInputInterface {
     void injectEvent(float x, float y, int action, int pointerId);
     void injectHoverEvent(float x, float y, int pointerId);
-    void injectScroll(float x, float y, int value);
+    void injectScroll(float x, float y, float value);
     void pauseResumeKeymap();
     KeymapConfig getKeymapConfig();
     KeyEventHandler getKeyEventHandler();
